Crate sdk

SDK for reading csv files efficiently from disk and providing command line argument parsing for data manipulation. Current functionality provides filtering of large csv files reducing the I/O overhead from high-level programming languages like Python

Panics

If the file path is not provided as a command line argument, a panic will result

Modules

data

Data structures and functions for working with data. This module contains the following submodules manipulation.

reader

Provide CsvMetadata struct and csv_reader function to read csv files

stdin_parser

Module argparser contains a public function parser

writer

Provides csv_writer function to write a csv file to disk

Functions

loader

Loader function that reads a csv file and returns a BufReader