sdk::data::manipulation::filtering
Function filter

pub fn filter(
    buffer: BufReader<File>,
    csv_struct: &CsvMetadata,
) -> Result<String, Error>

Safe data filtering function, single column & query matching

This function will take in a buffer reader, query and column name and return a vector of filtered lines as strings, where each line contains at least one field that matches the query.

Limitations

This function will have undesired results if the csv elements contain commas within double quotes For example, due to the limitation of the split function, the following line will be split into 4 fields 1,“a,b”,2,3 -> [1, “a, b”, 2, 3]

Panics

This function will panic if the column name is not found in the csv file