sdk::reader
Struct CsvMetadata

pub struct CsvMetadata {
    pub file: String,
    pub delimiter: char,
    pub has_header: bool,
    pub column_types: Vec<String>,
}

A reader module defines the struct containing metadata of the csv file

Example

Build a CsvMetadata struct

let metadata = CsvMetadata {
    file: "/path/to/file.csv".to_string(),
    delimiter: ',',
    has_header: true,
    column_types: vec!["string".to_string()]
};

Fields
file: Stringdelimiter: charhas_header: boolcolumn_types: Vec<String>
Implementations

impl CsvMetadata

pub fn clone(&self) -> CsvMetadata

Instanstiate self and return clone of the CsvMetadata struct

pub fn validate(&self) -> bool

Validate the csv file based on the file extension and headers for multiple files

The CsvMetadata struct is validated based on the file extension matching the *.csv extention. Additionally, for glob files the headers are validated to ensure that all files have the same headers. Otherwise, the concatenation of the files will result in a misaligned data structure. In other words, there will be shifts to the schema of the data.

Note

The CsvMetadata::validate function is a wrapper for two private functions validate_extension and validate_multifile_header If there are a large number of files, the validation process may take some time. This is caused by the need to read the headers of all files to check for consistency.